benchmark/pyproject.toml
[tool.poetry]
name = "ugc_2"
version = "0.1.0"
description = "ugc sprint 2"
authors = ["Sergey Bilibin"]
readme = "README.md"

[tool.poetry.dependencies]
elasticsearch = ">=8.7.0"
python = ">=3.10,<4.0"
python-dotenv="==1.0.0"
pydantic = "==2.1.1"
pydantic-settings = "^2.0.2"
pymongo = "^4.4.1"
uvicorn = {extras = ["standard"], version = "^0.23.1"}
typer = "^0.9.0"

[tool.autoflake]
in-place=true
expand-star-imports=true
remove-duplicate-keys=true
remove-all-unused-imports=true

[tool.black]
line-length = 120
skip-string-normalization=false

[tool.mypy]
disallow-untyped-defs=true
disallow-incomplete-defs=true
disallow-untyped-calls=true


[tool.isort]
multi_line_output=3
include_trailing_comma=true
force_grid_wrap=0
use_parentheses=true
line_length=120

[tool.flake8]
ignore = ['E203', 'E266', 'E501', 'W503']
max-line-length = 120
max-complexity = 18
select = ['B','C','E','F','W','T4']
exclude= [
    '.env.example',
    '.env',
    '.env.dev'
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"